# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PharmLink - Pharmacy Employment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # Candidates
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications (for candidate or employer depending on token)
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"

    post:
      operationId: createApplication
      tags: [applications]
      summary: Apply to a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "429":
          description: Quota exceeded
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      summary: Update application status (employer action)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # Offers
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - in: query
          name: role
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
...
  /admin/verify/{candidateId}:
    patch:
      operationId: verifyCandidateCredentials
      tags: [admin]
      summary: Verify or reject a candidate's credentials
      parameters:
        - in: path
          name: candidateId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateProfile"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [email, password, role, fullName]
      properties:
        email: { type: string }
        password: { type: string, minLength: 6 }
        role: { type: string, enum: [candidate, employer, admin] }
        fullName: { type: string }
        phone: { type: ["string", "null"] }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, role, fullName, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        role: { type: string, enum: [candidate, employer, admin] }
        fullName: { type: string }
        phone: { type: ["string", "null"] }
        createdAt: { type: string }

    # Candidate
    CandidateProfile:
      type: object
      required: [id, userId, fullName, governorate, city, specialization, yearsExperience, availableForWork, profileCompletionPct, isVerified, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        fullName: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        whatsapp: { type: ["string", "null"] }
        dateOfBirth: { type: ["string", "null"] }
        gender: { type: ["string", "null"] }
        governorate: { type: string }
        city: { type: string }
        address: { type: ["string", "null"] }
        university: { type: ["string", "null"] }
        faculty: { type: ["string", "null"] }
        graduationYear: { type: ["integer", "null"] }
        grade: { type: ["string", "null"] }
        specialization: { type: string }
        yearsExperience: { type: integer }
        previousWorkplaces: { type: ["string", "null"] }
        pharmacySkills: { type: ["string", "null"] }
        computerSkills: { type: ["string", "null"] }
        communicationSkills: { type: ["string", "null"] }
        certifications: { type: ["string", "null"] }
        courses: { type: ["string", "null"] }
        availableForWork: { type: boolean }
        preferredShift: { type: ["string", "null"] }
        jobType: { type: ["string", "null"] }
        maxDistanceKm: { type: ["integer", "null"] }
        profilePicture: { type: ["string", "null"] }
        personalSummary: { type: ["string", "null"] }
        cvUrl: { type: ["string", "null"] }
        lat: { type: ["number", "null"] }
        lng: { type: ["number", "null"] }
        profileCompletionPct: { type: integer }
        isVerified: { type: boolean }
        verificationBadge: { type: ["string", "null"] }
        createdAt: { type: string }
...
    Application:
      type: object
      required: [id, jobId, candidateId, status, createdAt]
      properties:
        id: { type: integer }
        jobId: { type: integer }
        candidateId: { type: integer }
        candidateName: { type: ["string", "null"] }
        jobTitle: { type: ["string", "null"] }
        employerName: { type: ["string", "null"] }
        coverLetter: { type: ["string", "null"] }
        status: { type: string, enum: [pending, reviewing, accepted, rejected] }
        createdAt: { type: string }
    ApplicationInput:
      type: object
      required: [jobId]
      properties:
        jobId: { type: integer }
        coverLetter: { type: string }
    ApplicationStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [reviewing, accepted, rejected] }

    # Offers
...
    AdminUserListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    VerificationInput:
      type: object
      required: [verified]
      properties:
        verified: { type: boolean }
        badge: { type: string }
        notes: { type: string }
...