# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmLink - Pharmacy Employment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  email: string;
  role: UserRole;
  fullName: string;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
}