# lib/api-zod/src/generated/types/verificationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PharmLink - Pharmacy Employment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface VerificationInput {
  verified: boolean;
  badge?: string;
  notes?: string;
}